testing now

+++ backend/src/services/tweet.py
@@ -15,32 +15,30 @@ from backend.src.utils.exception import CustomException
 
 
 class TweetService:
-
     @classmethod
-    async def add_tweet(cls,tweet:TweetWrite):
+    async def add_tweet(cls,tweet:TweetWrite, current_user_id: int) ->int:
         logger.debug('Добавление твита')
         async  with async_session() as db:
             new_tweet = TweetOrm(
-                data=tweet.data, author_id=tweet.author_id
+                data=tweet.data, author_id=current_user_id
             )
             # Добавляем в индекс, фиксируем, но не записываем в БД!!!
             db.add(new_tweet)
             await db.commit()
             await db.refresh(new_tweet)
 
-
             tweet_media_ids = tweet.tweet_media_ids
             if tweet_media_ids and tweet_media_ids != []:
                 # Привязываем изображения к твиту
                   await MediaService.update_media(tweet_media_ids=tweet_media_ids,tweet_id=new_tweet.id)
 
-
             # Сохраняем в БД все изменения (новый твит + привязку картинок к твиту)
 
             return new_tweet.id
 
+
     @classmethod
-    async def get_tweets(cls)->List[TweetRead]:
+    async def get_tweets(cls)->List[TweetOrm]:
         logger.debug('просмотр ленты твитов')
         async  with async_session() as db:
             query = (select(TweetOrm)
@@ -48,16 +46,15 @@ class TweetService:
                      joinedload(TweetOrm.author),
                      joinedload(TweetOrm.likes).subqueryload(LikeOrm.user),
                      joinedload(TweetOrm.media))
-                     .order_by(TweetOrm.created_at.desc())
+                     .order_by(TweetOrm.created_at)
                      )
 
             result =  await db.execute(query)
             return result.unique().scalars().all()
 
 
-
     @classmethod
-    async def get_tweet(cls,tweet_id) -> List[TweetRead]:
+    async def get_tweet(cls,tweet_id) -> TweetOrm|None:
             logger.debug(f"Поиск твита по id: {tweet_id}")
             async  with async_session() as db:
                 query = select(TweetOrm).where(TweetOrm.id == tweet_id)
@@ -65,14 +62,24 @@ class TweetService:
 
                 return tweet.scalar_one_or_none()
 
+
     @classmethod
-    async def delete_tweet(cls, id: int):
+    async def delete_tweet(cls,current_user_id:int, tweet_id: int)->None:
         logger.debug('удаление твита')
         async  with async_session() as db:
-            existing_tweet = await db.get(TweetOrm, id)
+            existing_tweet = await db.get(TweetOrm, tweet_id)
             if existing_tweet is None:
                 logger.error("Твит не найден")
                 raise CustomException(status_code=404, detail="not found")
 
-            await db.delete(existing_tweet)
-            await db.commit()
+            else:
+                if existing_tweet.author_id!=current_user_id:
+                     logger.warning("Ты удаляет чужой твит")
+                     raise CustomException(
+                            status_code=423,
+                            detail="The tweet that is being accessed is locked"
+                     )
+                else:
+
+                     await db.delete(existing_tweet)
+                     await db.commit()
